feat : P4 — 클라가 토너먼트 낙인을 직접 쓰지 않는다
TournamentProgress.MarkRewardPending 을 걷어냈다. 승리 낙인은 이제 서버
reportTournamentWin 이 세우고, 클라는 응답 채택으로 받는다. 이로써 tournament
슬롯에 클라가 쓰는 경로는 ResetForDebug 하나만 남는다(P0 가 가드로 감쌀 대상).

신고는 두 곳에서 쏜다. 전투 씬(BattleOutcome)이 먼저인 것은 기존 낙인이 거기
있던 이유와 같다 — 캐리어가 메모리라 로비까지 미루면 씬 로딩 중 종료가 승리를
삼킨다. 로비 복귀(TournamentReturnFlow)가 한 번 더 쏘는 것은 그 순간 네트워크가
없었을 때를 메우기 위해서다. 서버가 재신고를 AlreadyPending 으로 성공 처리하고
창구가 겹친 왕복을 합치므로 비용은 왕복 1회다.

선물 등장은 신고 왕복이 끝난 뒤로 미뤘다 — 낙인이 서기 전에 선물을 내면 눌러도
수령이 튕긴다. 실패해도 신호는 조건 없이 낸다: 그것이 맵의 등장 예약(정점을
재워 두는 것)을 푸는 유일한 열쇠라, 걸러 버리면 정점이 감춰진 채 남는다.

전투 씬 호출에 취소 토큰을 물리지 않는다. 씬 파괴가 업로드 봉인 해제
(InvokeAsync 의 finally) 전에 취소를 던지면 이후 저장이 통째로 막힌다.

재수화는 Tournament 슬롯만 넣었다(AlbumReward 는 P5 몫). 다른 슬롯과 달리 Init
계열이 아니라 통지뿐인데, TournamentProgress 가 세이브를 직독하고 캐시를 두지
않아 채택 시점에 값은 이미 새것이고 모르는 것은 화면뿐이기 때문이다.

표시용으로 격하한 판정에 주석을 남겼다(StateOf · CanEnter · IsChapterComplete)
— CardPackOpener.Precheck 와 같은 성격이고, 엇갈리면 서버가 이긴다.

requiredPoints 는 long 이다(RankGradeConfig.entryPoints 와 같은 축).
UserSaveData.VERSION 은 8 그대로 — 세이브 필드 변화가 0이다.

검증: Unity 컴파일 오류 0건. MarkRewardPending 잔여 호출부 0건.

+++ Assets/Scripts/Editor/SpecFirestoreUploader.Composition.cs
@@ -34,7 +34,7 @@ public static partial class SpecFirestoreUploader
         public string nodeId;
         public int order;
         public string prevNodeId;
-        public int requiredPoints;
+        public long requiredPoints;
     }
 
     /// <summary>CardAlbumConfig 저작을 AlbumEntry 표로 올린다. 성공하면 보고 줄을, 실패하면 null과 _error를 준다.</summary>
@@ -257,9 +257,9 @@ public static partial class SpecFirestoreUploader
     // 돌려주므로 points 0 인 신규 계정도 클라에선 첫 등급으로 읽힌다. entryPoints 를 그대로 쓰면
     // 서버만 그 계정을 잠근다.
     static bool TryBuildGradeEntryPoints(
-        RankConfig _config, out Dictionary<ERankGrade, int> _entryPoints, out string _error)
+        RankConfig _config, out Dictionary<ERankGrade, long> _entryPoints, out string _error)
     {
-        _entryPoints = new Dictionary<ERankGrade, int>();
+        _entryPoints = new Dictionary<ERankGrade, long>();
         _error = null;
 
         List<RankGradeConfig> t_grades = _config.grades;
@@ -310,7 +310,7 @@ public static partial class SpecFirestoreUploader
     {
         _rows = new List<TournamentChapterRow>();
 
-        if (!TryBuildGradeEntryPoints(_rankConfig, out Dictionary<ERankGrade, int> t_entryPoints, out _error))
+        if (!TryBuildGradeEntryPoints(_rankConfig, out Dictionary<ERankGrade, long> t_entryPoints, out _error))
             return false;
 
         var t_chapterIds = new HashSet<string>(StringComparer.Ordinal);
@@ -343,7 +343,7 @@ public static partial class SpecFirestoreUploader
                 return false;
             }
 
-            if (!t_entryPoints.TryGetValue(t_chapter.requiredGrade, out int t_requiredPoints))
+            if (!t_entryPoints.TryGetValue(t_chapter.requiredGrade, out long t_requiredPoints))
             {
                 _error = $"챕터 '{t_chapter.chapterId}' 의 requiredGrade '{t_chapter.requiredGrade}' 가 " +
                          "RankConfig.grades 에 없다 — 서버가 잠금을 잴 점수를 만들 수 없다.";
